rename yaml deps
