+++ test/ManuallySpecified/testdata/invalid-deps.yaml
